ai-news-agent/list_models.py
import requests
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("LLM_API_KEY")

url = f"https://generativelanguage.googleapis.com/v1beta/models?key={api_key}"
response = requests.get(url)
print(response.json())
