ai-news-agent/requirements.txt
feedparser==6.0.11
openai>=1.50.0
python-dotenv==1.0.1